% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topical_guide.R
\name{lds_datatable}
\alias{lds_datatable}
\title{lds_datatable}
\usage{
lds_datatable(x)
}
\arguments{
\item{x}{The output from `lds_topic()`}
}
\value{
a datatable from the library DT
}
\description{
Creates a DT::datatable object for a specific lds topic.
}
