% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topical_guide.R
\name{lds_expand_reference}
\alias{lds_expand_reference}
\title{lds_expand_reference}
\usage{
lds_expand_reference(reference = "1 Ne. 8:14–15, 18")
}
\arguments{
\item{reference}{a specific scripture reference as a character string. Should use the `verse_short_title`}
}
\value{
a character string with all text from the input verses
}
\description{
Function retrieves the verses of specific topic from the topical guide on lds.org
}
\details{
The function can handle `,`, `–`, or `-`. The `–` and `-` will pull all verses referenced.
}
\examples{
lds_expand_reference("1 Ne. 8:14–15, 18"); lds_expand_reference("1 Ne. 8:14-15, 18")
}
