% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topical_guide.R
\name{lds_scripture_url}
\alias{lds_scripture_url}
\title{lds_scripture_url}
\usage{
lds_scripture_url(references, text_link = TRUE)
}
\arguments{
\item{text_link}{A boolean that identifies if the output should be an html text link.}

\item{reference}{a specific scripture reference, or string of references, as a character string. Requires use of the `verse_short_title`}
}
\value{
a character string that is a url or text url.
}
\description{
Function creates url link to verse on lds.org
}
\examples{
lds_scripture_url("1 Ne. 8:14–15, 18"); lds_scripture_url("1 Ne. 8:14–15, 18", FALSE)
}
