% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topical_guide.R
\name{lds_topic}
\alias{lds_topic}
\title{lds_topic}
\usage{
lds_topic(topic)
}
\arguments{
\item{topic}{a correctly named topic from the topical guide}
}
\value{
a tibble with five columns - verse reference from the web, book, reference, text_summary, and full_text
}
\description{
Function retrieves the verses of specific topic from the topical guide on lds.org
}
\details{
the reference column matches the values in `scriptures$verse_short_title`
}
\examples{
lds_topic("God, Omniscience of")
}
