% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topical_guide.R
\name{lds_topics}
\alias{lds_topics}
\title{lds_topics}
\usage{
lds_topics(letter)
}
\arguments{
\item{letter}{a letter of the alphabet}
}
\description{
Function retrieves the list of topics from the topical guide on lds.org
}
\examples{
lds_topics("a")
}
