% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topical_guide.R
\name{lds_verse_join}
\alias{lds_verse_join}
\title{lds_verse_join}
\usage{
lds_verse_join(reference = "1 Ne. 8:14–15, 18")
}
\arguments{
\item{reference}{a specific scripture reference as a character string. Should use the `verse_short_title`}
}
\value{
a character string of the first short verse for merging.
}
\description{
Function pulls the first `scriptures$verse_short_title` element for merging with `scriptures`.
}
\details{
The function can handle `,`, `–`, or `-`. The `–` and `-` will keep the first verse reference.
}
\examples{
lds_verse_join("1 Ne. 8:14–15, 18"); lds_verse_join("1 Ne. 8:14-15, 18"); lds_verse_join("2 Ne. 2:3,5")
}
