% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topical_guide.R
\name{process_topics}
\alias{process_topics}
\title{process_topics}
\usage{
process_topics(topics)
}
\arguments{
\item{topics}{the ouput from within the `lds_topics` function.}
}
\value{
a tibble with three columns - 'site_text', 'topic', and 'see'
}
\description{
process topics data using the lds_topics function
}
\details{
by using the url 'https://www.lds.org/scriptures/tg/' the value in the 'topic' column can be appended to get to the page with the scripture references for that topic.
}
